# Fix traces of decoding error for tuples